Cleanup code

+++ BudgedManager/Controllers/ExpenseController.cs
@@ -1,215 +1,175 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
+using BudgedManager.Models;
+using BudgedManager.Models.Entity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using BudgedManager.Models;
-using BudgedManager.Models.Entity;
 
-namespace BudgedManager.Controllers
+namespace BudgedManager.Controllers;
+
+public class ExpenseController : Controller
 {
-    public class ExpenseController : Controller
+    private readonly AppDbContext _context;
+
+    public ExpenseController(AppDbContext context)
     {
-        private readonly AppDbContext _context;
+        _context = context;
+    }
 
-        public ExpenseController(AppDbContext context)
+    // GET: Expense
+    public async Task<IActionResult> Index(string? orderBy, string? date, string? category, string? amount)
+    {
+        var expenses = await _context.Expenses.Include(
+                e => e.Category)
+            .ToListAsync();
+        if (date != null)
+            /*
+             * OPTIONAL: Date.ToString("d") creates example: "11/12/2024"
+             */
+            expenses = new List<Expense>(expenses.Where(e => e.Date.Date == DateTime.Parse(date)));
+        if (category != null) expenses = new List<Expense>(expenses.Where(s => s.Category.Name.Equals(category)));
+        if (amount != null) expenses = new List<Expense>(expenses.Where(s => s.Amount.ToString() == amount));
+        switch (orderBy)
         {
-            _context = context;
+            case "Amount":
+                expenses = new List<Expense>(expenses.OrderBy(e => e.Amount));
+                break;
+            case "Category":
+                expenses = new List<Expense>(expenses.OrderBy(e => e.Category.Name));
+                break;
+            case "Date":
+                expenses = new List<Expense>(expenses.OrderBy(e => e.Date));
+                break;
         }
 
-        // GET: Expense
-        public async Task<IActionResult> Index(string? orderBy, string? date, string? category, string? amount)
+        return View(expenses);
+    }
+
+
+    // GET: Expense/Summary/
+    public async Task<IActionResult> Summary(DateTime? startDate, DateTime? endDate)
+    {
+        var categoryExpenses = _context.Expenses
+            .GroupBy(e => e.CategoryId)
+            .Select(group => new SummaryDto
+            {
+                CategoryId = group.Key,
+                CategoryName = group.First().Category.Name,
+                Score = (float)Math.Round(group.Sum(e => e.Amount))
+            }).ToList();
+
+        var countRecords = await _context.Expenses.Select(e => e.Date.Date).Distinct().CountAsync();
+
+        var totalExpenses = await _context.Expenses.SumAsync(e => e.Amount);
+
+        var averageDays = totalExpenses / countRecords;
+        // var averageWeeks = totalExpenses / (countRecords / 7);
+        // var averageMonths = totalExpenses / (countRecords / 30);
+
+        ViewData["Sum"] = (float)Math.Round(totalExpenses, 2);
+        ViewData["Highest"] = (float)Math.Round(categoryExpenses.Max(e => e.Score), 2);
+        ViewData["categoryExpenses"] = categoryExpenses;
+        ViewData["AverageDays"] = (float)Math.Round(averageDays, 2);
+        // ViewData["AverageWeeks"] = (float)Math.Round(averageWeeks,2) ;
+        // ViewData["AverageMonths"] = (float)Math.Round(averageMonths,2) ;
+
+        return View();
+    }
+
+    // GET: Expense/Details/5
+    public async Task<IActionResult> Details(int? id)
+    {
+        if (id == null) return NotFound();
+
+        var expense = await _context.Expenses
+            .FirstOrDefaultAsync(m => m.Id == id);
+        if (expense == null) return NotFound();
+
+        return View(expense);
+    }
+
+    // GET: Expense/Create
+    public IActionResult Create()
+    {
+        ViewData["Category"] = new SelectList(_context.Categories, "Id", "Name");
+        return View();
+    }
+
+    // POST: Expense/Create
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Create([Bind("Id,Amount,CategoryId,Date,Comment")] Expense expense)
+    {
+        if (expense.CategoryId.Equals(null)) return BadRequest();
+
+        if (ModelState.IsValid)
         {
-            var expenses = await _context.Expenses.Include(
-                    e => e.Category)
-                .ToListAsync();
-            //todo try to optimize that process 
-            if (date != null)
-            {
-                /*
-                 * OPTIONAL: Date.ToString("d") creates example: "11/12/2024"
-                 */
-                expenses = new List<Expense>(expenses.Where(e => e.Date.Date == DateTime.Parse(date)));
-            }
-            if (category != null)
-            {
-                expenses = new List<Expense>(expenses.Where(s => s.Category.Name.Equals(category)));
-            }
-            if (amount != null)
-            {
-                expenses = new List<Expense>(expenses.Where(s => s.Amount.ToString() == amount));
-            }
-            switch (orderBy)
-            {
-                case "Amount":
-                    expenses = new List<Expense>(expenses.OrderBy(e => e.Amount));
-                    break;
-                case "Category":
-                    expenses = new List<Expense>(expenses.OrderBy(e => e.Category.Name));
-                    break;
-                case "Date":
-                    expenses = new List<Expense>(expenses.OrderBy(e => e.Date));
-                    break;
-            }
-            return View(expenses);
+            expense.Amount = decimal.Parse(string.Format("{0:0.00}", expense.Amount));
+            _context.Add(expense);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
+        return View(expense);
+    }
 
-        // GET: Expense/Summary/
-        public async Task<IActionResult> Summary(DateTime? startDate, DateTime? endDate)
+    // GET: Expense/Edit/5
+    public async Task<IActionResult> Edit(int? id)
+    {
+        if (id == null) return NotFound();
+        ViewData["Category"] = new SelectList(_context.Categories, "Id", "Name");
+
+        var expense = await _context.Expenses.FindAsync(id);
+        if (expense == null) return NotFound();
+        return View(expense);
+    }
+
+    // POST: Expense/Edit/5
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Edit(int id, [Bind("Id,Amount,Category,Date,Comment")] Expense expense)
+    {
+        if (id != expense.Id) return NotFound();
+
+        if (ModelState.IsValid)
         {
-            var categoryExpenses = _context.Expenses
-                .GroupBy(e => e.CategoryId)
-                .Select(group => new SummaryDto
-                {
-                    CategoryId = group.Key,
-                    CategoryName = group.First().Category.Name,
-                    Score = (float)Math.Round(group.Sum(e => e.Amount))
-                }).ToList();
-
-            var countRecords = await _context.Expenses.Select(e => e.Date.Date).Distinct().CountAsync();
-            
-            var totalExpenses = await _context.Expenses.SumAsync(e => e.Amount);
-
-            var averageDays = totalExpenses / countRecords;
-            // var averageWeeks = totalExpenses / (countRecords / 7);
-            // var averageMonths = totalExpenses / (countRecords / 30);
-            
-            ViewData["Sum"] = (float)Math.Round(totalExpenses,2) ;
-            ViewData["Highest"] = (float)Math.Round(categoryExpenses.Max(e => e.Score),2) ;
-            ViewData["categoryExpenses"] = categoryExpenses;
-            ViewData["AverageDays"] = (float)Math.Round(averageDays,2) ;
-            // ViewData["AverageWeeks"] = (float)Math.Round(averageWeeks,2) ;
-            // ViewData["AverageMonths"] = (float)Math.Round(averageMonths,2) ;
-            
-            return View();
-        }
-
-        // GET: Expense/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
+            try
             {
-                return NotFound();
-            }
-
-            var expense = await _context.Expenses
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
-
-            return View(expense);
-        }
-
-        // GET: Expense/Create
-        public IActionResult Create()
-        {
-            ViewData["Category"] = new SelectList(_context.Categories, "Id", "Name");
-            return View();
-        }
-
-        // POST: Expense/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Amount,CategoryId,Date,Comment")] Expense expense)
-        {
-            if (expense.CategoryId.Equals(null))
-            {
-                return BadRequest();
-            }
-            
-            if (ModelState.IsValid)
-            {
-                expense.Amount = decimal.Parse(String.Format("{0:0.00}", expense.Amount));
-                _context.Add(expense);
+                _context.Update(expense);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
             }
-            
-            return View(expense);
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!ExpenseExists(expense.Id)) return NotFound();
+
+                throw;
+            }
+
+            return RedirectToAction(nameof(Index));
         }
 
-        // GET: Expense/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-            ViewData["Category"] = new SelectList(_context.Categories, "Id", "Name");
-            
-            var expense = await _context.Expenses.FindAsync(id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
-            return View(expense);
-        }
+        return View(expense);
+    }
 
-        // POST: Expense/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Amount,Category,Date,Comment")] Expense expense)
-        {
-            if (id != expense.Id)
-            {
-                return NotFound();
-            }
+    // GET: Expense/Delete/5
+    public async Task<IActionResult> Delete(int? id)
+    {
+        if (id == null) return NotFound();
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(expense);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ExpenseExists(expense.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(expense);
-        }
+        var expense = await _context.Expenses
+            .FirstOrDefaultAsync(m => m.Id == id);
+        if (expense == null) return NotFound();
 
-        // GET: Expense/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
+        return View(expense);
+    }
 
-            var expense = await _context.Expenses
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
-
-            return View(expense);
-        }
-        // GET: Expense/History
-        public async Task<IActionResult> History()
-        {
-            var expense = _context.Expenses.Select(
+    // GET: Expense/History
+    public async Task<IActionResult> History()
+    {
+        var expense = _context.Expenses.Select(
                 group =>
                     new HistoryDto
                     {
@@ -218,30 +178,27 @@ namespace BudgedManager.Controllers
                         Date = group.Date,
                         Comment = group.Comment
                     })
-                .OrderByDescending(a => a.Date)
-                .ToList();
-            ViewData["expenses"] = expense;
-            return View();
-        }
+            .OrderByDescending(a => a.Date)
+            .ToList();
+        ViewData["expenses"] = expense;
+        return View();
+    }
 
-        // POST: Expense/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var expense = await _context.Expenses.FindAsync(id);
-            if (expense != null)
-            {
-                _context.Expenses.Remove(expense);
-            }
+    // POST: Expense/Delete/5
+    [HttpPost]
+    [ActionName("Delete")]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> DeleteConfirmed(int id)
+    {
+        var expense = await _context.Expenses.FindAsync(id);
+        if (expense != null) _context.Expenses.Remove(expense);
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+        await _context.SaveChangesAsync();
+        return RedirectToAction(nameof(Index));
+    }
 
-        private bool ExpenseExists(int id)
-        {
-            return _context.Expenses.Any(e => e.Id == id);
-        }
+    private bool ExpenseExists(int id)
+    {
+        return _context.Expenses.Any(e => e.Id == id);
     }
 }
